requirements.txt
PyPDF2==3.0.1
python-pptx==0.6.23
requests==2.31.0
python-dotenv==1.0.0
